* ParseAligns/ParseAligns.cpp (percent): Fix this compiler warning: passing `double' for converting 1 of `std::_Setw std::setw(int)'
git-svn-id: https://svn.bcgsc.ca/svn/abyss/trunk@59168 ad02020f-2a26-4133-947e-1829317c1416

+++ ParseAligns/ParseAligns.cpp
@@ -458,7 +458,7 @@ static void readAlignmentsFile(string path, ReadAlignMap* pout)
 static string percent(size_t x, size_t n)
 {
 	ostringstream ss;
-	ss << setw(log10(n) + 1) << x;
+	ss << setw((int)log10(n) + 1) << x;
 	if (x > 0)
 		ss << "  " << setprecision(3) << (float)100*x/n << '%';
 	return ss.str();
